Force clean only if -f specified.

+++ portainer_install.sh
@@ -6,6 +6,9 @@
 sudo apt remove apparmor
 sudo docker stop portainer
 sudo docker rm portainer
-sudo docker volume rm portainer_data
-sudo docker volume create portainer_data
+
+if [[ "$1" == "-f" || "$1" == "--force" ]]; then
+  sudo docker volume rm portainer_data
+  sudo docker volume create portainer_data
+fi
 sudo docker run -d -p 8000:8000 -p 9443:9443 --name portainer --restart=always -v /var/run/docker.sock:/var/run/docker.sock -v portainer_data:/data portainer/portainer-ee:latest
